Move build ellipsis function
